make file-mapped array implementations able to read files larger than 2GB

+++ src/main/java/net/sci/array/numeric/impl/FileMappedFloat32Array3D.java
@@ -76,7 +76,7 @@ public class FileMappedFloat32Array3D extends Float32Array3D
         this.filePath = filePath;
 
         // initialize slice offset, assuming there are fully packed
-        int sliceItemCount = size0 * size1 * 4;
+        long sliceItemCount = size0 * size1 * 4;
         this.offsets = new long[size2];
         for (int z = 0; z < size2; z++)
         {
